move chat context files to context/use-chat/
