+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { NoMessagesGreeting } from "@/components/a1/empty-states/no-messages";
 import { MainChatInput } from "@/components/a1/input/main-chat-input";
 import { MessageParts } from "@/components/a1/messages";
-import { useChatMessages } from "@/contexts/chat-hooks";
+import { useChatMessages } from "@/contexts/use-chat/chat-hooks";
 import { cn } from "@/lib/utils";
 
 function App() {
